config saving (partially) implemented

+++ source/json.h
@@ -2,6 +2,6 @@
 #define JSON_H
 
 int jsonLoad(char* path, s16 selCount, char** selections, u8* selectArray);
-int jsonSave(char* path, int selCount, char** selections, u8* selectArray);
+int jsonSave(char* path, s16 selCount, char** selections, u8* selectArray);
 
 #endif
